# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Pipeline Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type DealStage = (typeof DealStage)[keyof typeof DealStage];

export const DealStage = {
  lead: "lead",
  qualified: "qualified",
  proposal: "proposal",
  negotiation: "negotiation",
  closed_won: "closed_won",
  closed_lost: "closed_lost",
} as const;

export interface Owner {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl: string | null;
  createdAt: string;
}

export interface CreateOwnerBody {
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface UpdateOwnerBody {
  name?: string;
  email?: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface Deal {
  id: number;
  title: string;
  companyName: string;
  stage: DealStage;
  value: number;
  expectedCloseDate: string;
  ownerId: number;
  /** @nullable */
  ownerName: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDealBody {
  title: string;
  companyName: string;
  stage: DealStage;
  value: number;
  expectedCloseDate: string;
  ownerId: number;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateDealBody {
  title?: string;
  companyName?: string;
  stage?: DealStage;
  value?: number;
  expectedCloseDate?: string;
  ownerId?: number;
  /** @nullable */
  notes?: string | null;
}

export interface PipelineSummary {
  totalDeals: number;
  totalValue: number;
  wonValue: number;
  lostValue: number;
  activeDeals: number;
  activeValue: number;
  avgDealValue: number;
  winRate: number;
}

export interface StageBreakdown {
  stage: DealStage;
  count: number;
  totalValue: number;
}

export interface OwnerBreakdown {
  ownerId: number;
  ownerName: string;
  count: number;
  totalValue: number;
}

export type ListDealsParams = {
  stage?: DealStage;
  ownerId?: number;
};