# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Pipeline Tracker API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all deal owners
 */
export const ListOwnersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  avatarUrl: zod.string().nullable(),
  createdAt: zod.coerce.date(),
});
export const ListOwnersResponse = zod.array(ListOwnersResponseItem);

/**
 * @summary Create a new deal owner
 */
export const CreateOwnerBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  avatarUrl: zod.string().nullish(),
});

/**
 * @summary Get a deal owner by ID
 */
export const GetOwnerParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOwnerResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  avatarUrl: zod.string().nullable(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update a deal owner
 */
export const UpdateOwnerParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateOwnerBody = zod.object({
  name: zod.string().optional(),
  email: zod.string().optional(),
  avatarUrl: zod.string().nullish(),
});

export const UpdateOwnerResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  avatarUrl: zod.string().nullable(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a deal owner
 */
export const DeleteOwnerParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all deals
 */
export const ListDealsQueryParams = zod.object({
  stage: zod
    .enum([
      "lead",
      "qualified",
      "proposal",
      "negotiation",
      "closed_won",
      "closed_lost",
    ])
    .optional(),
  ownerId: zod.coerce.number().optional(),
});

export const ListDealsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  companyName: zod.string(),
  stage: zod.enum([
    "lead",
    "qualified",
    "proposal",
    "negotiation",
    "closed_won",
    "closed_lost",
  ]),
  value: zod.number(),
  expectedCloseDate: zod.coerce.date(),
  ownerId: zod.number(),
  ownerName: zod.string().nullable(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListDealsResponse = zod.array(ListDealsResponseItem);

/**
 * @summary Create a new deal
 */
export const CreateDealBody = zod.object({
  title: zod.string(),
  companyName: zod.string(),
  stage: zod.enum([
    "lead",
    "qualified",
    "proposal",
    "negotiation",
    "closed_won",
    "closed_lost",
  ]),
  value: zod.number(),
  expectedCloseDate: zod.coerce.date(),
  ownerId: zod.number(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get a deal by ID
 */
export const GetDealParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDealResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  companyName: zod.string(),
  stage: zod.enum([
    "lead",
    "qualified",
    "proposal",
    "negotiation",
    "closed_won",
    "closed_lost",
  ]),
  value: zod.number(),
  expectedCloseDate: zod.coerce.date(),
  ownerId: zod.number(),
  ownerName: zod.string().nullable(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a deal
 */
export const UpdateDealParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateDealBody = zod.object({
  title: zod.string().optional(),
  companyName: zod.string().optional(),
  stage: zod
    .enum([
      "lead",
      "qualified",
      "proposal",
      "negotiation",
      "closed_won",
      "closed_lost",
    ])
    .optional(),
  value: zod.number().optional(),
  expectedCloseDate: zod.coerce.date().optional(),
  ownerId: zod.number().optional(),
  notes: zod.string().nullish(),
});

export const UpdateDealResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  companyName: zod.string(),
  stage: zod.enum([
    "lead",
    "qualified",
    "proposal",
    "negotiation",
    "closed_won",
    "closed_lost",
  ]),
  value: zod.number(),
  expectedCloseDate: zod.coerce.date(),
  ownerId: zod.number(),
  ownerName: zod.string().nullable(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a deal
 */
export const DeleteDealParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get pipeline summary statistics
 */
export const GetPipelineSummaryResponse = zod.object({
  totalDeals: zod.number(),
  totalValue: zod.number(),
  wonValue: zod.number(),
  lostValue: zod.number(),
  activeDeals: zod.number(),
  activeValue: zod.number(),
  avgDealValue: zod.number(),
  winRate: zod.number(),
});

/**
 * @summary Get deal counts and values grouped by stage
 */
export const GetDealsByStageResponseItem = zod.object({
  stage: zod.enum([
    "lead",
    "qualified",
    "proposal",
    "negotiation",
    "closed_won",
    "closed_lost",
  ]),
  count: zod.number(),
  totalValue: zod.number(),
});
export const GetDealsByStageResponse = zod.array(GetDealsByStageResponseItem);

/**
 * @summary Get deal counts and values grouped by owner
 */
export const GetDealsByOwnerResponseItem = zod.object({
  ownerId: zod.number(),
  ownerName: zod.string(),
  count: zod.number(),
  totalValue: zod.number(),
});
export const GetDealsByOwnerResponse = zod.array(GetDealsByOwnerResponseItem);

/**
 * @summary Get deals closing within the next 30 days
 */
export const GetClosingSoonDealsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  companyName: zod.string(),
  stage: zod.enum([
    "lead",
    "qualified",
    "proposal",
    "negotiation",
    "closed_won",
    "closed_lost",
  ]),
  value: zod.number(),
  expectedCloseDate: zod.coerce.date(),
  ownerId: zod.number(),
  ownerName: zod.string().nullable(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetClosingSoonDealsResponse = zod.array(
  GetClosingSoonDealsResponseItem,
);