# lib/api-zod/src/generated/types/deal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Pipeline Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { DealStage } from "./dealStage";

export interface Deal {
  id: number;
  title: string;
  companyName: string;
  stage: DealStage;
  value: number;
  expectedCloseDate: Date;
  ownerId: number;
  /** @nullable */
  ownerName: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: Date;
  updatedAt: Date;
}