# lib/api-zod/src/generated/types/dealStage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Pipeline Tracker API
 * OpenAPI spec version: 0.1.0
 */

export type DealStage = (typeof DealStage)[keyof typeof DealStage];

export const DealStage = {
  lead: "lead",
  qualified: "qualified",
  proposal: "proposal",
  negotiation: "negotiation",
  closed_won: "closed_won",
  closed_lost: "closed_lost",
} as const;