# lib/api-zod/src/generated/types/ownerBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Pipeline Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface OwnerBreakdown {
  ownerId: number;
  ownerName: string;
  count: number;
  totalValue: number;
}