# lib/api-zod/src/generated/types/pipelineSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Pipeline Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface PipelineSummary {
  totalDeals: number;
  totalValue: number;
  wonValue: number;
  lostValue: number;
  activeDeals: number;
  activeValue: number;
  avgDealValue: number;
  winRate: number;
}