# lib/api-zod/src/generated/types/stageBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Pipeline Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { DealStage } from "./dealStage";

export interface StageBreakdown {
  stage: DealStage;
  count: number;
  totalValue: number;
}