# lib/api-zod/src/generated/types/updateOwnerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Pipeline Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateOwnerBody {
  name?: string;
  email?: string;
  /** @nullable */
  avatarUrl?: string | null;
}